Set default build var for custom config file to `None`
Previously the role was using an empty string to indicate lack of a
custom config file, but this caused unexpected behavior when used in
tandem with Ansible's `with_first_found` loop construct. Specifically,
the empty string caused Ansible to interpret the target file as the
`tasks/` directory within the role, and it subsequently tried to copy up
all YAML task lists within that directory. Not what we want!

Using `None` is a saner default, and causes `with_first_found` to behave
sanely again.