Commits latest kernel config used for SecureDrop
Allows us to publish the current kernel config for Ubuntu Trusty, used
for SecureDrop, for transparency and accountability. The config is
extremely simple: default Ubuntu Trusty settings, with the grsecurity
high-security options enabled on top.

Updates the Ansible playbook for the SecureDrop kernel build process to
import the new config automatically. This will still need to change
if/when we migrate to the 4.x kernel series.