Cleans up GPG key fetching tasks
The logic is slightly modified from the implementation developed during
review of #100. Hadn't sufficiently tested the logic there prior to
merge, so removed an extra in-line "item" that was causing the tasks to
fail, and updated the README with docs on the new format.

The new GPG pubkey addition was required for the head of the Ubuntu
Trusty kernel overlay.