Creates dedicated `grsec-rebuild-securedrop` VM
For SecureDrop we build and retain linux-source packages, which include
a tarball of the entire kernel source immediately prior to configuration
and compilation. Anyone trying to use that source tarball to rebuild a
kernel image package will need a few things:

  * build environment (apt dependencies and such)
  * tarball of source (requested from FPF)
  * ubuntu-overlay

The Ubuntu Trusty kernel overlay is *not* included in the source
tarball, so we much clone that separately. The recommended command in
the rebuild playbook explicitly states that the overlay should be
applied via CLI flags.

In order to achieve reuse of the grsec build logic, without attempting
to fetch the restricted-access grsecurity patches, added a few tags to
the main task list so that the tasks can be selectively included. That
way, anyone who wants to rebuild the kernel image from source does not
need authentication credentials for grsecurity. Instead, everything
required will be provided by the playbook and docs.

Also tweaks the `grsec-install` machine to use the same Trusty base
image as all the other Trusty boxes, for consistency in testing.