Uses MIT PGP keyserver by default
The role needs to fetch GPG keys in order to verify signatures on the
downloaded files. The previous default keyserver used an HKPS pool, and
was terribly unreliable: sometimes key lookups would fail with cryptic
messages, then on a subsequent run of the playbook, everything was
happy. That's frustrating, especially in the context of building
kernels, since a kicked off task should run for several hours, and a
failed playbook can waste a lot of time.

Let's fall back to plain and simple pgp.mit.edu. If someone feels
strongly about this choice, it's still a var, so they're welcome to
override it. The new default setting is much more reliable when running
multiple iterations of the config.