Prunes Stretch packages
The `kernel-package` package didn't make it into Debian Stretch,
although it was in Jessie (and will be in Buster, too). We'll have
to rearchitect the package vars. We don't actually need kernel-package,
just a few utilities such as `bc` and `xz`, as long as we're willing
to forego use of `make-kpkg` in favor of `make deb-pkg`.