Configures normal user in container image
The role logic already fails when run as root, and Docker containers run
as root by default (:face_with_rolling_eyes:). So let's configure a
normal user in the Dockerfile and force it via USER. In order to elevate
privileges to install packages and such, we'll need sudo—so let's set up
passwordless sudo as part of the testing image.