Creates separate "unofficial" unpacking task flow
Newcomers to the role are unlikely to have access to the grsecurity
subscription program. Part of the motivation for this role is to
encourage system adminstrators to explore grsecurity, so they may opt to
subscribe, but we'd do well to catch folks before that point, as well.

Back when the grsecurity beta patches were public (no longer the
case[0]), newcomers could use the beta/test patches to evaluate
the grsecurity patch and feature set. Helping contributors understand
the grsecurity kernel hardening is critical for improving the kernel
maintenance story for SecureDrop [1], and the paywall has proven a major
barrier to increasing the community's understanding of these hardening
measures.

So let's try out the Debian team's unofficial backports of the
grsecurity beta patches. They're not appropriate for use in production
SecureDrop, but they will allow full end-to-end builds, which will
enable contributors to recompile with different options, and to explore
the world of PaX flag management for Python-heavy applications.

Most of the control flow for the role is managed via the
`tasks/main*.yml` files, as per normal. The unpacking logic, in which
the tarballs are verified and extracted under the "stable" flow, cannot
easily be reused for the unofficial flow, so let's simply write another
task list.

Hardcodes many of the path names, since we cannot reuse the
`grsecurity_urls.py` fact-gathering module.

The bulk of the logic showing up in the diff is simply moved or
borrowed, but there are two genuine changes to the logic:

    * Sets default patch type to "unofficial"
    * Uses Debian Stretch "backports" repo for access to patch

Everything else remains largely the same.

[0] https://grsecurity.net/passing_the_baton.php
[1] https://github.com/freedomofpress/securedrop