Creates Molecule scenario for unofficial builds
Reusing the logic from the official story, as far as Docker containers
go, and playbook too, largely. The vars are different, mainly the lack
of overrides, but otherwise we've got a clean story here.

Plucked out the "revision" var in the default compile command, since it
was dependent on vars fetched from the `grsecurity_urls` module, and
that module is not used in the unofficial flow.

Uses a bundled config file simply so builds can be unattended. It's
rather minimal, but I'm able to run the image produced from it fine on
a ThinkPad T470. \m/