Updates README with Molecule docs
The most cursory of explanations, simply to mention that Molecule is
used for testing and provide an example command. Future PRs will improve
the Molecule testing story, e.g. with containers, so not going nuts with
the explanations now, since the entrypoints will likely change soon.

The documented scenario is SecureDrop-specific, since that's a common
use case for inspecting this role.