Removes unnecessary test steps in unofficial scenario
Since the role is building packages, and therefore changing state on
every run, we'll never pass the idempotence check. We're not using the
prepare step, either. Let's also ditch the linting and verify steps for
now, since there's a lot of cleanup to do there across all scenarios.

Many of the ansible-lint recommended changes are not appropriate for
this role, so we can either use "noqa" throughout or just skip linting.