Defaults to make targets, rather than make-kpkg
Sets make deb-pkg as default compile command, still overrideable via a
var. Also using `make clean` instead of `make-kpkg clean`, since
make-kpkg will likely not be available on many platforms.

Adjusts the SecureDrop build process to use the new logic, as well.

Updates fetch pattern to target src tarballs

The switch from `make-kpkg` to `make deb-pkg` means that the kernel
source is bottled up in a tarball now, not a deb package as before.
In order to maintain GPL compliance and ensure that distributed binaries
are backed with sources upon request, we'll fetch back the source
tarball for safekeeping on the Ansible controller.

Updates README with new default vars, so the role is accessible to
newcomers.