Provide option to patch static kernel config file
This is useful in certain scenarios where we want to build a kernel
based off an existing config file but also apply slightly different
values during compilation. It doesnt always make sense to keep
maintaining lists of static configs for each set of tweaks. This commit
allows one to pass in an ansible variable file filled with config
directives.