Fix error on Ubuntu when attempting to build XEN
This resolved an issued I was seeing where the `make deb-pkg` was
bombing out thanks to this missing dependency. It's a fairly small
package so figured I'd add it to the baseline. If we continue to need
more for this, I'll introduce a new vars files and tweak the logic so it
gets mixed in with the existing package list.