Sets 4.4 config as default for SecureDrop builds
The "securedrop" Molecule scenario is used for managing the SecureDrop
kernel build process. We've been tinkering with the config options for
some time now, and we're finally stabilizing into something that works
well in Trusty. Will get this up into a testing apt repo for further QA
with VMs, ramping up to release.

Note that all future builds will continue to use this config. Revert
this commit in order to resort to manual configuration.